fixed duplicate code and lib.ts dep
