+++ src/lib/adjudicators.ts
@@ -1,8 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: 2023-Present The Pepr Authors
 
-import { Event } from "./types";
-import { Operation } from "./mutate-types";
+import { Event, Operation } from "./enums";
 import {
   __,
   allPass,
