+++ src/lib/filter.ts
@@ -2,7 +2,7 @@
 // SPDX-FileCopyrightText: 2023-Present The Pepr Authors
 
 import { AdmissionRequest, Binding } from "./types";
-import { Operation } from "./mutate-types";
+import { Operation } from "./enums";
 import {
   carriesIgnoredNamespace,
   carriedName,
