+++ src/lib/validate-request.ts
@@ -8,7 +8,7 @@ import { KubernetesObject } from "kubernetes-fluent-client";
 import { clone } from "ramda";
 import { AdmissionRequest } from "./types";
 import { ValidateActionResponse } from "./types";
-import { Operation } from "./mutate-types";
+import { Operation } from "./enums";
 
 /**
  * The RequestWrapper class provides methods to modify Kubernetes objects in the context
